Drain interrupts before compaction
theap may want to evacuate some objects, and it uses the VM interrupt to
do so.  Draining the interrupts ensures that no objects that get moved
will be impacted by theap.  Before this change, an object could be added
for evacuation, but compacting the heap might move the object which
makes the pointer go bad in the theap queue.