Set FL_EXIVAR when we have to mutate the global object_id table
Removing objects from the global id table is expensive.  This commit
sets FL_EXIVAR on objects that have their object id set in the global
table.  That way we know to only try removal when the flag is set.
Freeing generic ivars isn't cheap anyway, so this should be a pretty
rare case.